✨ support cancellation
